tree-wide: Use named semaphores for better debuggability.
Named semaphore is an enhancement to the usual seastar::semaphore, which
is initialized with a name string that helps to identify which semaphore
it is; when the semaphore raises an exception (i.e. timed out, broken),
the name string is included in the error message.

Tree-wide commit, for two reasons:

1. We need to use named semaphores widely to get the debuggability benefit.
2. The coupling between the semaphore_units<> type and the named
   semaphore specialization causes the choice of semaphore to leak quite a
   bit, making a treewide change easier.

Fixes: #5489