cluster: do not put ephemeral creds in a snapshot
Previously, the security manager was inserting ephemeral credentials
into the snapshot that is reloaded on controller log replay. This is a
problem because, upon log replay, RP will reload an ephemeral credential
that has a username and password but no principal. This leads to
authorization failures.

Therefore, this commit stops the security manager from insterting
ephemeral credentials into a snapshot.

Fixes #11141