dt/kafka: quotas test without bursting
Currently, when the client quota tests try to produce under the client
quota limit, they are actually producing slightly over the limit because
while their produced data is under the limit, their full request size
including the header is over the limit.

The reason why the test are currently passing is because the current
implementation allows bursting up to 10 windows worth.

We want to not allow this bursting behaviour with the new
token-bucket-based implementation and so the tests are fixed here to
produce 8 messages instead of 10 to stay within the configured limit.