pandaproxy/sr: Refactor store internal state to use unparsed schemas
- Replace canonical schemas for unparsed schemas stored in store and
move all schema processing to sharded_store. This allows to have
lazy schema validation on startup and relax requirements for valid
schema. With this changes, a schema that had been imported before its
dependencies will be found valid when it is requested.

- Add backport to sharded_store to be able to access an unparsed schema
as-is, without processing it. This is needed as now there might be an
invalid schema loaded at startup. Operations like
delete_subject_version, which needs to reference a schema before
deleting it would be problematic.