cluster: add `topic_table::get_name_by_id(...)`
Implement an efficient lookup for finding a topic by its topic id. This
will be used by kafka handlers to translate the topic id received on a
request to its topic name.

The optional topic id is always checked for nullopt-ness, since the code
needs to work during the migration period when not all topics have topic
ids assigned.

This is implemented using a wrapper class `underlying_map` that wraps
most const methods of the topics map, and only provides mutable access
through the `mutate(...)` method, which ensures that the invariants are
maintained, ie. that the topic id mapping remains in sync.